Devuan: setup initramfs-tools and u-boot for LUKS via dropbear ssh in initrd
To unlock LUKS wait a few seconds after plugging the USB Armory in
(the LED will be lit in low power), then
  ssh unlock@10.0.0.1  # with password "unlock" without quotes
You will be asked automatically to type in the key for the LUKS partition.

Original crypt_unlock.sh and fix-login.sh are from
http://hacksr.blogspot.co.at/2012/05/ssh-unlock-with-fully-encrypted-ubuntu.html
and copied and modified with permission granted via private email from
Christoph Gritschenberger.