Add Release Pipeline
A Github Action will be triggered when creating a Release that will use the new
make target for release to build and compress the assets.  Once successful, the
assets are uploaded to the release that triggered the action.

REQUIREMENTS:
 * create a Personal Access Token with full Repo scope permissions
 * create a GitHub Secret in the  repo named: GITHUB_TOKEN_FOR_RELEASES

Changes to code involved:

 * New make target for release
 * Renaming the assets to include the emmc or usd targets in the filename
 * Updates to README.md

Note that I did run into some timeouts and files-too-large issues (without
compressing).  The xz compression handles too large files; however, if we run
into timeouts, just Re-run Release action, which should kick off another build
and only upload assets that did not exist.

A note on "already existing": The github action I am using does not exit
gracefully nor handle re-uploads - it fails the entire build.  I'll fork the
action and fix it for future revisions.